First cut at picking specific units for conversion
Use the UI state passed down to FSM search engine to construct a subset
of unit-recognizer only for those units that are checked.

There's a bit of extra complexity going on here to support default
values. In a fresh install the persisted UI state is empty (as in empty
object), so the code tries to compensate for that by treating the
missing properties as set to true (current default state in html).

The 'nomen: true' option in Gruntfile gags lint about leading underscore
in function names.